+++ src/main/java/handler/dynamic/MyPageDynamicHtmlHandler.java
@@ -1,4 +1,4 @@
-package handler.dynamic_handler;
+package handler.dynamic;
 
 import db.Database;
 import http.cookie.Cookie;
@@ -9,22 +9,20 @@ import http.response.HttpResponse;
 import http.session.SessionManager;
 import model.User;
 
-import java.nio.charset.StandardCharsets;
-
-public class WriteDynamicHtmlHandler implements DynamicHtmlHandler{
+public class MyPageDynamicHtmlHandler implements  DynamicHtmlHandler{
     private final SessionManager sessionManager = SessionManager.getInstance();
     private static final String USERNAME = "<!-- USERNAME -->";
 
     @Override
     public HttpResponse handle(byte[] fileData, String extension, HttpRequest httpRequest) {
-        String htmlContent = new String(fileData, StandardCharsets.UTF_8);
+        String htmlContent = new String(fileData);
 
         Cookie cookie = httpRequest.getCookie("sessionId");
 
         if(cookie == null){
             return new HttpResponse.Builder()
-                    .httpStatus(HttpStatus.FOUND)
-                    .location("http://localhost:8080/login")
+                    .httpStatus(HttpStatus.TEMPORARY_REDIRECT)
+                    .location("http://localhost:8080/")
                     .build();
         }
 
@@ -42,4 +40,6 @@ public class WriteDynamicHtmlHandler implements DynamicHtmlHandler{
         User user = Database.findUserById(userId);
         return user.getName();
     }
+
+
 }
