+++ src/main/java/handler/dynamic/WriteDynamicHtmlHandler.java
@@ -1,4 +1,4 @@
-package handler.dynamic_handler;
+package handler.dynamic;
 
 import db.Database;
 import http.cookie.Cookie;
@@ -9,20 +9,22 @@ import http.response.HttpResponse;
 import http.session.SessionManager;
 import model.User;
 
-public class MyPageDynamicHtmlHandler implements  DynamicHtmlHandler{
+import java.nio.charset.StandardCharsets;
+
+public class WriteDynamicHtmlHandler implements DynamicHtmlHandler{
     private final SessionManager sessionManager = SessionManager.getInstance();
     private static final String USERNAME = "<!-- USERNAME -->";
 
     @Override
     public HttpResponse handle(byte[] fileData, String extension, HttpRequest httpRequest) {
-        String htmlContent = new String(fileData);
+        String htmlContent = new String(fileData, StandardCharsets.UTF_8);
 
         Cookie cookie = httpRequest.getCookie("sessionId");
 
         if(cookie == null){
             return new HttpResponse.Builder()
-                    .httpStatus(HttpStatus.TEMPORARY_REDIRECT)
-                    .location("http://localhost:8080/")
+                    .httpStatus(HttpStatus.FOUND)
+                    .location("http://localhost:8080/login")
                     .build();
         }
 
@@ -40,6 +42,4 @@ public class MyPageDynamicHtmlHandler implements  DynamicHtmlHandler{
         User user = Database.findUserById(userId);
         return user.getName();
     }
-
-
 }
