+++ src/main/java/handler/request/FileRequestHandler.java
@@ -1,4 +1,4 @@
-package handler.request_handler;
+package handler.request;
 
 import handler.mapping.DynamicHtmlHandlerMapping;
 import http.request.HttpRequest;
